+++ src/Elmish.DragAndDrop/DropArea.fs
@@ -15,9 +15,10 @@ module DropArea =
       | None ->
         let children =
           content
-          |> List.map (fun (elementId : ElementId, gen : ElementGenerator) -> 
-            renderDraggable DragStatus.NoActiveDrag model config elementId dispatch gen
+          |> List.map (fun (elementId, gen) ->
+            toDraggables DragStatus.NoActiveDrag model config elementId dispatch gen
           )
+          |> List.concat
         div props children
       | Some { StartLocation = (listIndex, index, draggedElementId) } ->
         let props = [
@@ -28,9 +29,35 @@ module DropArea =
         let children =
           content
           |> List.map (fun (elementId, gen) ->
-            renderDraggable (DragStatus.ActiveDrag draggedElementId) model config elementId dispatch gen
+            toDraggables (DragStatus.ActiveDrag draggedElementId) model config elementId dispatch gen
           )
+          |> List.concat
+
         div props children
+    
+    static member fromGeneratorsWithTag model dispatch config (props: IHTMLProp list) content tag =
+      match model.Moving with
+      | None ->
+        let children =
+          content
+          |> List.map (fun (elementId, gen) ->
+            toDraggables DragStatus.NoActiveDrag model config elementId dispatch gen
+          )
+          |> List.concat
+        tag (Seq.ofList props) (Seq.ofList children)
+      | Some { StartLocation = (listIndex, index, draggedElementId) } ->
+        let props = [
+          yield! props
+          Listeners.defaultReleaseListener dispatch
+          Listeners.defaultMouseMoveListener dispatch
+        ]
+        let children =
+          content
+          |> List.map (fun (elementId, gen) ->
+            toDraggables (DragStatus.ActiveDrag draggedElementId) model config elementId dispatch gen
+          )
+          |> List.concat
+        tag (Seq.ofList props) (Seq.ofList children)
 
     static member fromDragHandles model dispatch config props content =
       match model.Moving with