+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Allow the user to specify the HTML tag to use to render a draggable and a drop area. This allows Tables to be used.
+
 ## [0.2.0-rc2] - 2021-05-22
 
 ### Added